Front End Food Categories
Prepared a fixed list of food categories that can be used to run queries depending on the food type desired. Added navigation between the sign up screen and the food type screen. In the future, we should redirect a user form the login screen to the main screen once the login has been set up properly.